+++ backend/src/main/java/com/duzz/backend/controller/SubjectController.java
@@ -1,5 +1,6 @@
 package com.duzz.backend.controller;
 
+import com.duzz.backend.component.SecurityProvider;
 import com.duzz.backend.form.SubjectAddForm;
 import com.duzz.backend.service.SubjectService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -15,11 +16,13 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class SubjectController {
     private final SubjectService subjectService;
+    private final SecurityProvider securityProvider;
 
     @PostMapping("/add")
     @Operation(summary = "과목을 추가합니다.")
     public ResponseEntity<?> addSubject(@RequestBody SubjectAddForm form) {
         try {
+            securityProvider.checkAdmin();
             var subjectDto = SubjectAddForm.toDto(form);
             subjectService.addOrUpdateSubject(subjectDto);
             return ResponseEntity.ok("과목 추가 성공");
@@ -33,6 +36,7 @@ public class SubjectController {
     @Operation(summary = "과목을 삭제합니다.")
     public ResponseEntity<?> deleteSubject(@PathVariable String subjectId) {
         try {
+            securityProvider.checkAdmin();
             subjectService.deleteSubject(subjectId);
             return ResponseEntity.ok("과목 삭제 성공");
         } catch (Exception e) {
